backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

psycopg2-binary==2.9.9

bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6

#validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

#cors
python-jose==3.3.0